♻️ Reconstruct the app layer and service layer to comply with the MVC three-layer architecture.
